+++ hs-src/Language/Scheme/Primitives.hs
@@ -47,6 +47,8 @@ module Language.Scheme.Primitives (
  , hashTblValues 
  , hashTblCopy
  , hashTblMake
+ , wrapHashTbl
+ , wrapLeadObj
  -- ** String
  , buildString
  , makeString
@@ -91,6 +93,7 @@ module Language.Scheme.Primitives (
  , unpackEquals 
  , boolBinop 
  , unaryOp 
+ , unaryOp'
  , strBoolBinop 
  , charBoolBinop 
  , boolBoolBinop
@@ -123,6 +126,7 @@ module Language.Scheme.Primitives (
 import Language.Scheme.Numerical
 import Language.Scheme.Parser
 import Language.Scheme.Types
+import Language.Scheme.Variables
 --import qualified Control.Exception
 import Control.Monad.Error
 import qualified Data.ByteString as BS
@@ -150,6 +154,7 @@ try' = tryIOError
 
 makePort :: IOMode -> [LispVal] -> IOThrowsError LispVal
 makePort mode [String filename] = liftM Port $ liftIO $ openFile filename mode
+makePort mode [p@(Pointer _ _)] = recDerefPtrs p >>= box >>= makePort mode
 makePort _ [] = throwError $ NumArgs (Just 1) []
 makePort _ args@(_ : _) = throwError $ NumArgs (Just 1) args
 
@@ -211,9 +216,12 @@ readCharProc _ args@(_ : _) = throwError $ BadSpecialForm "" $ List args
 {- writeProc :: --forall a (m :: * -> *).
              (MonadIO m, MonadError LispError m) =>
              (Handle -> LispVal -> IO a) -> [LispVal] -> m LispVal -}
-writeProc func [obj] = writeProc func [obj, Port stdout]
+writeProc func [obj] = do
+    dobj <- recDerefPtrs obj -- Last opportunity to do this before writing
+    writeProc func [dobj, Port stdout]
 writeProc func [obj, Port port] = do
-    output <- liftIO $ try' (liftIO $ func port obj)
+    dobj <- recDerefPtrs obj -- Last opportunity to do this before writing
+    output <- liftIO $ try' (liftIO $ func port dobj)
     case output of
         Left _ -> throwError $ Default "I/O error writing to port"
         Right _ -> return $ Nil ""
@@ -233,12 +241,14 @@ writeCharProc other = if length other == 2
                      else throwError $ NumArgs (Just 2) other
 
 fileExists, deleteFile :: [LispVal] -> IOThrowsError LispVal
+fileExists [p@(Pointer _ _)] = recDerefPtrs p >>= box >>= fileExists
 fileExists [String filename] = do
     exists <- liftIO $ doesFileExist filename
     return $ Bool exists
 fileExists [] = throwError $ NumArgs (Just 1) []
 fileExists args@(_ : _) = throwError $ NumArgs (Just 1) args
 
+deleteFile [p@(Pointer _ _)] = recDerefPtrs p >>= box >>= deleteFile
 deleteFile [String filename] = do
     output <- liftIO $ try' (liftIO $ removeFile filename)
     case output of
@@ -249,6 +259,7 @@ deleteFile args@(_ : _) = throwError $ NumArgs (Just 1) args
 
 readContents :: [LispVal] -> IOThrowsError LispVal
 readContents [String filename] = liftM String $ liftIO $ readFile filename
+readContents [p@(Pointer _ _)] = recDerefPtrs p >>= box >>= readContents
 readContents [] = throwError $ NumArgs (Just 1) []
 readContents args@(_ : _) = throwError $ NumArgs (Just 1) args
 
@@ -260,6 +271,7 @@ load filename = do
      else throwError $ Default $ "File does not exist: " ++ filename
 
 readAll :: [LispVal] -> IOThrowsError LispVal
+readAll [p@(Pointer _ _)] = recDerefPtrs p >>= box >>= readAll
 readAll [String filename] = liftM List $ load filename
 readAll [] = throwError $ NumArgs (Just 1) []
 readAll args@(_ : _) = throwError $ NumArgs (Just 1) args
@@ -273,6 +285,7 @@ _gensym prefix = do
 -- |Generate a (reasonably) unique symbol, given an optional prefix.
 --  This function is provided even though it is not part of R5RS.
 gensym :: [LispVal] -> IOThrowsError LispVal
+gensym [p@(Pointer _ _)] = recDerefPtrs p >>= box >>= gensym
 gensym [String prefix] = _gensym prefix
 gensym [] = _gensym " g"
 gensym args@(_ : _) = throwError $ NumArgs (Just 1) args
@@ -283,20 +296,25 @@ gensym args@(_ : _) = throwError $ NumArgs (Just 1) args
 ---------------------------------------------------
 
 -- List primitives
-car :: [LispVal] -> ThrowsError LispVal
+car :: [LispVal] -> IOThrowsError LispVal
+car [p@(Pointer _ _)] = derefPtr p >>= box >>= car
 car [List (x : _)] = return x
 car [DottedList (x : _) _] = return x
 car [badArg] = throwError $ TypeMismatch "pair" badArg
 car badArgList = throwError $ NumArgs (Just 1) badArgList
 
-cdr :: [LispVal] -> ThrowsError LispVal
+cdr :: [LispVal] -> IOThrowsError LispVal
+cdr [p@(Pointer _ _)] = derefPtr p >>= box >>= cdr
 cdr [List (_ : xs)] = return $ List xs
 cdr [DottedList [_] x] = return x
 cdr [DottedList (_ : xs) x] = return $ DottedList xs x
 cdr [badArg] = throwError $ TypeMismatch "pair" badArg
 cdr badArgList = throwError $ NumArgs (Just 1) badArgList
 
-cons :: [LispVal] -> ThrowsError LispVal
+cons :: [LispVal] -> IOThrowsError LispVal
+cons [x, p@(Pointer _ _)] = do
+  y <- derefPtr p
+  cons [x, y]
 cons [x1, List []] = return $ List [x1]
 cons [x, List xs] = return $ List $ x : xs
 cons [x, DottedList xs xlast] = return $ DottedList (x : xs) xlast
@@ -324,8 +342,7 @@ makeVector [(Number n), a] = do
 makeVector [badType] = throwError $ TypeMismatch "integer" badType
 makeVector badArgList = throwError $ NumArgs (Just 1) badArgList
 
-buildVector (o : os) = do
-  let lst = o : os
+buildVector lst@(o : os) = do
   return $ Vector $ (listArray (0, length lst - 1)) lst
 buildVector badArgList = throwError $ NumArgs (Just 1) badArgList
 
@@ -350,7 +367,7 @@ listToVector [badType] = throwError $ TypeMismatch "list" badType
 listToVector badArgList = throwError $ NumArgs (Just 1) badArgList
 
 -- ------------ Bytevector Primitives --------------
-makeByteVector, byteVector, byteVectorLength, byteVectorRef, byteVectorCopy, byteVectorAppend, byteVectorUtf2Str, byteVectorStr2Utf :: [LispVal] -> ThrowsError LispVal
+makeByteVector, byteVector :: [LispVal] -> ThrowsError LispVal
 makeByteVector [(Number n)] = do
   let ls = replicate (fromInteger n) (0 :: Word8)
   return $ ByteVector $ BS.pack ls
@@ -366,6 +383,10 @@ byteVector bs = do
    conv (Number n) = fromInteger n :: Word8
    conv n = 0 :: Word8
 
+byteVectorLength, byteVectorRef, byteVectorCopy, byteVectorAppend, byteVectorUtf2Str :: [LispVal] -> IOThrowsError LispVal
+byteVectorCopy (p@(Pointer _ _) : lvs) = do
+    bv <- derefPtr p
+    byteVectorCopy (bv : lvs)
 byteVectorCopy [ByteVector bv] = do
     return $ ByteVector $ BS.copy
         bv
@@ -384,16 +405,24 @@ byteVectorCopy badArgList = throwError $ NumArgs (Just 1) badArgList
 
 byteVectorAppend bs = do
     let acc = BS.pack []
-        conv (ByteVector bs) = bs
-        conv x = BS.empty
-        bs' = map conv bs
+        conv :: LispVal -> IOThrowsError BSU.ByteString
+        conv p@(Pointer _ _) = do
+          bs <- derefPtr p
+          conv bs
+        conv (ByteVector bs) = return bs
+        conv x = return BS.empty
+    bs' <- mapM conv bs
     return $ ByteVector $ BS.concat bs'
 -- TODO: error handling
 
+byteVectorLength [p@(Pointer _ _)] = derefPtr p >>= box >>= byteVectorLength
 byteVectorLength [(ByteVector bv)] = return $ Number $ toInteger $ BS.length bv
 byteVectorLength [badType] = throwError $ TypeMismatch "bytevector" badType
 byteVectorLength badArgList = throwError $ NumArgs (Just 1) badArgList
 
+byteVectorRef (p@(Pointer _ _) : lvs) = do
+    bv <- derefPtr p
+    byteVectorRef (bv : lvs)
 byteVectorRef [(ByteVector bv), (Number n)] = do
     let len = toInteger $ (BS.length bv) - 1
     if n > len || n < 0
@@ -402,17 +431,41 @@ byteVectorRef [(ByteVector bv), (Number n)] = do
 byteVectorRef [badType] = throwError $ TypeMismatch "bytevector integer" badType
 byteVectorRef badArgList = throwError $ NumArgs (Just 2) badArgList
 
+byteVectorUtf2Str [p@(Pointer _ _)] = derefPtr p >>= box >>= byteVectorUtf2Str
 byteVectorUtf2Str [(ByteVector bv)] = do
     return $ String $ BSU.toString bv 
 -- TODO: need to support other overloads of this function
 byteVectorUtf2Str [badType] = throwError $ TypeMismatch "bytevector" badType
 byteVectorUtf2Str badArgList = throwError $ NumArgs (Just 1) badArgList
+
+byteVectorStr2Utf :: [LispVal] -> IOThrowsError LispVal
+byteVectorStr2Utf [p@(Pointer _ _)] = derefPtr p >>= box >>= byteVectorStr2Utf
 byteVectorStr2Utf [(String s)] = do
     return $ ByteVector $ BSU.fromString s
 -- TODO: need to support other overloads of this function
 byteVectorStr2Utf [badType] = throwError $ TypeMismatch "string" badType
 byteVectorStr2Utf badArgList = throwError $ NumArgs (Just 1) badArgList
 
+
+-- ------------ Ptr Helper Primitives --------------
+
+wrapHashTbl, wrapLeadObj :: ([LispVal] -> ThrowsError LispVal) -> [LispVal] -> IOThrowsError LispVal
+wrapHashTbl fnc [p@(Pointer _ _)] = do
+  val <- derefPtr p
+  liftThrows $ fnc [val]
+wrapHashTbl fnc (p@(Pointer _ _) : key : args) = do
+  ht <- derefPtr p
+  k <- recDerefPtrs key
+  liftThrows $ fnc (ht : k : args)
+wrapHashTbl fnc args = liftThrows $ fnc args
+
+wrapLeadObj fnc [p@(Pointer _ _)] = do
+  val <- derefPtr p
+  liftThrows $ fnc [val]
+wrapLeadObj fnc (p@(Pointer _ _) : args) = do
+  obj <- derefPtr p
+  liftThrows $ fnc (obj : args)
+wrapLeadObj fnc args = liftThrows $ fnc args
 
 -- ------------ Hash Table Primitives --------------
 
@@ -490,17 +543,22 @@ doMakeString n char s =
        then String s
        else doMakeString (n - 1) char (s ++ [char])
 
-stringLength :: [LispVal] -> ThrowsError LispVal
+stringLength :: [LispVal] -> IOThrowsError LispVal
+stringLength [p@(Pointer _ _)] = derefPtr p  >>= box >>= stringLength
 stringLength [String s] = return $ Number $ foldr (const (+ 1)) 0 s -- Could probably do 'length s' instead...
 stringLength [badType] = throwError $ TypeMismatch "string" badType
 stringLength badArgList = throwError $ NumArgs (Just 1) badArgList
 
-stringRef :: [LispVal] -> ThrowsError LispVal
+stringRef :: [LispVal] -> IOThrowsError LispVal
+stringRef [p@(Pointer _ _)] = derefPtr p >>= box >>= stringRef
 stringRef [(String s), (Number k)] = return $ Char $ s !! fromInteger k
 stringRef [badType] = throwError $ TypeMismatch "string number" badType
 stringRef badArgList = throwError $ NumArgs (Just 2) badArgList
 
-substring :: [LispVal] -> ThrowsError LispVal
+substring :: [LispVal] -> IOThrowsError LispVal
+substring (p@(Pointer _ _) : lvs) = do
+  s <- derefPtr p
+  substring (s : lvs)
 substring [(String s), (Number start), (Number end)] =
   do let slength = fromInteger $ end - start
      let begin = fromInteger start
@@ -508,31 +566,42 @@ substring [(String s), (Number start), (Number end)] =
 substring [badType] = throwError $ TypeMismatch "string number number" badType
 substring badArgList = throwError $ NumArgs (Just 3) badArgList
 
-stringCIEquals :: [LispVal] -> ThrowsError LispVal
-stringCIEquals [(String str1), (String str2)] = do
-  if (length str1) /= (length str2)
-     then return $ Bool False
-     else return $ Bool $ ciCmp str1 str2 0
-  where ciCmp s1 s2 idx = if idx == (length s1)
-                             then True
-                             else if (toLower $ s1 !! idx) == (toLower $ s2 !! idx)
-                                     then ciCmp s1 s2 (idx + 1)
-                                     else False
-stringCIEquals [badType] = throwError $ TypeMismatch "string string" badType
-stringCIEquals badArgList = throwError $ NumArgs (Just 2) badArgList
+stringCIEquals :: [LispVal] -> IOThrowsError LispVal
+stringCIEquals args = do
+  List dargs <- recDerefPtrs $ List args
+  case dargs of
+    [(String str1), (String str2)] -> do
+      if (length str1) /= (length str2)
+         then return $ Bool False
+         else return $ Bool $ ciCmp str1 str2 0
+    [badType] -> throwError $ TypeMismatch "string string" badType
+    badArgList -> throwError $ NumArgs (Just 2) badArgList
+ where ciCmp s1 s2 idx = 
+         if idx == (length s1)
+            then True
+            else if (toLower $ s1 !! idx) == (toLower $ s2 !! idx)
+                    then ciCmp s1 s2 (idx + 1)
+                    else False
 
-stringCIBoolBinop :: ([Char] -> [Char] -> Bool) -> [LispVal] -> ThrowsError LispVal
-stringCIBoolBinop op [(String s1), (String s2)] = boolBinop unpackStr op [(String $ strToLower s1), (String $ strToLower s2)]
+stringCIBoolBinop :: ([Char] -> [Char] -> Bool) -> [LispVal] -> IOThrowsError LispVal
+stringCIBoolBinop op args = do 
+  List dargs <- recDerefPtrs $ List args -- Deref any pointers
+  case dargs of
+    [(String s1), (String s2)] ->
+      liftThrows $ boolBinop unpackStr op [(String $ strToLower s1), (String $ strToLower s2)]
+    [badType] -> throwError $ TypeMismatch "string string" badType
+    badArgList -> throwError $ NumArgs (Just 2) badArgList
   where strToLower str = map (toLower) str
-stringCIBoolBinop _ [badType] = throwError $ TypeMismatch "string string" badType
-stringCIBoolBinop _ badArgList = throwError $ NumArgs (Just 2) badArgList
 
 charCIBoolBinop :: (Char -> Char -> Bool) -> [LispVal] -> ThrowsError LispVal
 charCIBoolBinop op [(Char s1), (Char s2)] = boolBinop unpackChar op [(Char $ toLower s1), (Char $ toLower s2)]
 charCIBoolBinop _ [badType] = throwError $ TypeMismatch "character character" badType
 charCIBoolBinop _ badArgList = throwError $ NumArgs (Just 2) badArgList
 
-stringAppend :: [LispVal] -> ThrowsError LispVal
+stringAppend :: [LispVal] -> IOThrowsError LispVal
+stringAppend (p@(Pointer _ _) : lvs) = do
+  s <- derefPtr p
+  stringAppend (s : lvs)
 stringAppend [(String s)] = return $ String s -- Needed for "last" string value
 stringAppend (String st : sts) = do
   rest <- stringAppend sts
@@ -542,9 +611,12 @@ stringAppend (String st : sts) = do
 stringAppend [badType] = throwError $ TypeMismatch "string" badType
 stringAppend badArgList = throwError $ NumArgs (Just 1) badArgList
 
-stringToNumber :: [LispVal] -> ThrowsError LispVal
+stringToNumber :: [LispVal] -> IOThrowsError LispVal
+stringToNumber (p@(Pointer _ _) : lvs) = do
+  s <- derefPtr p
+  stringToNumber (s : lvs)
 stringToNumber [(String s)] = do
-  result <- (readExpr s)
+  result <- liftThrows $ readExpr s
   case result of
     n@(Number _) -> return n
     n@(Rational _) -> return n
@@ -561,24 +633,28 @@ stringToNumber [(String s), Number radix] = do
 stringToNumber [badType] = throwError $ TypeMismatch "string" badType
 stringToNumber badArgList = throwError $ NumArgs (Just 1) badArgList
 
-stringToList :: [LispVal] -> ThrowsError LispVal
+stringToList :: [LispVal] -> IOThrowsError LispVal
+stringToList [p@(Pointer _ _)] = derefPtr p >>= box >>= stringToList
 stringToList [(String s)] = return $ List $ map (Char) s
 stringToList [badType] = throwError $ TypeMismatch "string" badType
 stringToList badArgList = throwError $ NumArgs (Just 1) badArgList
 
-listToString :: [LispVal] -> ThrowsError LispVal
+listToString :: [LispVal] -> IOThrowsError LispVal
+listToString [p@(Pointer _ _)] = derefPtr p >>= box >>= listToString
 listToString [(List [])] = return $ String ""
-listToString [(List l)] = buildString l
+listToString [(List l)] = liftThrows $ buildString l
 listToString [badType] = throwError $ TypeMismatch "list" badType
 listToString [] = throwError $ NumArgs (Just 1) []
 listToString args@(_ : _) = throwError $ NumArgs (Just 1) args
 
-stringCopy :: [LispVal] -> ThrowsError LispVal
+stringCopy :: [LispVal] -> IOThrowsError LispVal
+stringCopy [p@(Pointer _ _)] = derefPtr p >>= box >>= stringCopy
 stringCopy [String s] = return $ String s
 stringCopy [badType] = throwError $ TypeMismatch "string" badType
 stringCopy badArgList = throwError $ NumArgs (Just 2) badArgList
 
-isDottedList :: [LispVal] -> ThrowsError LispVal
+isDottedList :: [LispVal] -> IOThrowsError LispVal
+isDottedList ([p@(Pointer _ _)]) = derefPtr p >>= box >>= isDottedList
 isDottedList ([DottedList _ _]) = return $ Bool True
 -- Must include lists as well since they are made up of 'chains' of pairs
 isDottedList ([List []]) = return $ Bool False
@@ -593,17 +669,19 @@ isProcedure ([IOFunc _]) = return $ Bool True
 isProcedure ([EvalFunc _]) = return $ Bool True
 isProcedure _ = return $ Bool False
 
-isVector, isList :: LispVal -> ThrowsError LispVal
+isVector,isByteVector, isList :: LispVal -> IOThrowsError LispVal
+isVector p@(Pointer _ _) = derefPtr p >>= isVector
 isVector (Vector _) = return $ Bool True
 isVector _ = return $ Bool False
+isByteVector p@(Pointer _ _) = derefPtr p >>= isVector
+isByteVector (ByteVector _) = return $ Bool True
+isByteVector _ = return $ Bool False
+isList p@(Pointer _ _) = derefPtr p >>= isList
 isList (List _) = return $ Bool True
 isList _ = return $ Bool False
 
-isByteVector :: LispVal -> ThrowsError LispVal
-isByteVector (ByteVector _) = return $ Bool True
-isByteVector _ = return $ Bool False
-
-isNull :: [LispVal] -> ThrowsError LispVal
+isNull :: [LispVal] -> IOThrowsError LispVal
+isNull ([p@(Pointer _ _)]) = derefPtr p >>= box >>= isNull
 isNull ([List []]) = return $ Bool True
 isNull _ = return $ Bool False
 
@@ -621,7 +699,8 @@ symbol2String [notAtom] = throwError $ TypeMismatch "symbol" notAtom
 symbol2String [] = throwError $ NumArgs (Just 1) []
 symbol2String args@(_ : _) = throwError $ NumArgs (Just 1) args
 
-string2Symbol :: [LispVal] -> ThrowsError LispVal
+string2Symbol :: [LispVal] -> IOThrowsError LispVal
+string2Symbol ([p@(Pointer _ _)]) = derefPtr p >>= box >>= string2Symbol
 string2Symbol ([String s]) = return $ Atom s
 string2Symbol [] = throwError $ NumArgs (Just 1) []
 string2Symbol [notString] = throwError $ TypeMismatch "string" notString
@@ -652,7 +731,8 @@ isChar :: [LispVal] -> ThrowsError LispVal
 isChar ([Char _]) = return $ Bool True
 isChar _ = return $ Bool False
 
-isString :: [LispVal] -> ThrowsError LispVal
+isString :: [LispVal] -> IOThrowsError LispVal
+isString [p@(Pointer _ _)] = derefPtr p >>= box >>= isString
 isString ([String _]) = return $ Bool True
 isString _ = return $ Bool False
 
@@ -683,10 +763,16 @@ unaryOp f [v] = f v
 unaryOp _ [] = throwError $ NumArgs (Just 1) []
 unaryOp _ args@(_ : _) = throwError $ NumArgs (Just 1) args
 
-{- numBoolBinop :: (Integer -> Integer -> Bool) -> [LispVal] -> ThrowsError LispVal
-numBoolBinop = boolBinop unpackNum -}
-strBoolBinop :: (String -> String -> Bool) -> [LispVal] -> ThrowsError LispVal
-strBoolBinop = boolBinop unpackStr
+unaryOp' :: (LispVal -> IOThrowsError LispVal) -> [LispVal] -> IOThrowsError LispVal
+unaryOp' f [v] = f v
+unaryOp' _ [] = throwError $ NumArgs (Just 1) []
+unaryOp' _ args@(_ : _) = throwError $ NumArgs (Just 1) args
+
+strBoolBinop :: (String -> String -> Bool) -> [LispVal] -> IOThrowsError LispVal
+strBoolBinop fnc args = do
+  List dargs <- recDerefPtrs $ List args -- Deref any pointers
+  liftThrows $ boolBinop unpackStr fnc dargs
+
 charBoolBinop = boolBinop unpackChar
 boolBoolBinop :: (Bool -> Bool -> Bool) -> [LispVal] -> ThrowsError LispVal
 boolBoolBinop = boolBinop unpackBool