+++ hs-src/Language/Scheme/Types.hs
@@ -22,11 +22,7 @@ module Language.Scheme.Types
     , LispError (..)
     , ThrowsError 
     , IOThrowsError 
-    , trapError
-    , extractValue 
     , liftThrows 
-    , runIOThrowsREPL 
-    , runIOThrows 
     -- * Types and related functions
     , LispVal (
           Atom
@@ -54,6 +50,7 @@ module Language.Scheme.Types
              , hbody
              , hclosure
         , IOFunc
+        , CustFunc
         , EvalFunc
         , Pointer
              , pointerVar
@@ -86,6 +83,7 @@ module Language.Scheme.Types
     , eqv 
     , eqvList
     , eqVal 
+    , box
     , makeFunc
     , makeNormalFunc
     , makeVarargs
@@ -161,17 +159,6 @@ instance Error LispError where
 -- |Container used by operations that could throw an error
 type ThrowsError = Either LispError
 
--- |Error handler that returns a string description of any error
-trapError :: -- forall (m :: * -> *) e.
-            (MonadError e m, Show e) =>
-             m String -> m String
-trapError action = catchError action (return . show)
-
--- |Utility function to unwrap a value from ThrowsError
-extractValue :: ThrowsError a -> a
-extractValue (Right val) = val
-extractValue (Left _) = error "Unexpected error in extractValue"
-
 -- |Container used to provide error handling in the IO monad
 type IOThrowsError = ErrorT LispError IO
 
@@ -179,20 +166,6 @@ type IOThrowsError = ErrorT LispError IO
 liftThrows :: ThrowsError a -> IOThrowsError a
 liftThrows (Left err) = throwError err
 liftThrows (Right val) = return val
-
--- |Execute an IO action and return result or an error message.
---  This is intended for use by a REPL, where a result is always
---  needed regardless of type.
-runIOThrowsREPL :: IOThrowsError String -> IO String
-runIOThrowsREPL action = runErrorT (trapError action) >>= return . extractValue
-
--- |Execute an IO action and return error or Nothing if no error was thrown.
-runIOThrows :: IOThrowsError String -> IO (Maybe String)
-runIOThrows action = do
-    runState <- runErrorT action
-    case runState of
-        Left err -> return $ Just (show err)
-        Right _ -> return $ Nothing
 
 -- |Scheme data types
 data LispVal = Atom String
@@ -236,7 +209,7 @@ data LispVal = Atom String
          body :: [LispVal],
          closure :: Env
         }
- -- ^Function
+ -- ^Function written in Scheme
  | HFunc {hparams :: [String],
           hvararg :: (Maybe String),
           hbody :: (Env -> LispVal -> LispVal -> Maybe [LispVal] -> IOThrowsError LispVal),
@@ -248,6 +221,10 @@ data LispVal = Atom String
  | EvalFunc ([LispVal] -> IOThrowsError LispVal)
  {- ^Function within the IO monad with access to
  the current environment and continuation. -}
+ | CustFunc ([LispVal] -> IOThrowsError LispVal)
+ -- ^A custom function written by code outside of husk.
+ --  Any code that uses the Haskell API should define custom
+ --  functions using this data type.
  | Pointer { pointerVar :: String
             ,pointerEnv :: Env } 
  -- ^Pointer to an environment variable.
@@ -412,6 +389,7 @@ eqv [x@(HFunc _ _ _ _), y@(Func _ _ _ _)] = do
 --
 eqv [x@(PrimitiveFunc _), y@(PrimitiveFunc _)] = return $ Bool $ (show x) == (show y)
 eqv [x@(IOFunc _), y@(IOFunc _)] = return $ Bool $ (show x) == (show y)
+eqv [x@(CustFunc _), y@(CustFunc _)] = return $ Bool $ (show x) == (show y)
 eqv [x@(EvalFunc _), y@(EvalFunc _)] = return $ Bool $ (show x) == (show y)
 -- FUTURE: comparison of two continuations
 eqv [l1@(List _), l2@(List _)] = eqvList eqv [l1, l2]
@@ -475,9 +453,14 @@ showVal (HFunc {hparams = args, hvararg = varargs, hbody = _, hclosure = _}) =
       Just arg -> " . " ++ arg) ++ ") ...)"
 showVal (Port _) = "<IO port>"
 showVal (IOFunc _) = "<IO primitive>"
+showVal (CustFunc _) = "<custom primitive>"
 showVal (EvalFunc _) = "<procedure>"
 showVal (Pointer p _) = "<ptr " ++ p ++ ">"
 showVal (Opaque d) = "<Haskell " ++ show (dynTypeRep d) ++ ">"
+
+-- |A helper function to make pointer deref code more concise
+box :: LispVal -> IOThrowsError [LispVal]
+box a = return [a]
 
 -- |Convert a list of Lisp objects into a space-separated string
 unwordsList :: [LispVal] -> String