+++ ChangeLog.markdown
@@ -4,6 +4,14 @@ v3.7.1
 - TBD - map/for-each multiple list args
 - Cache macro expansions to allow for a significant performance improvement
 
+OPT branch -  (TODO: need to revise all this. these are just rough notes for what is going on with this branch)
+
+Internally, husk uses a Pointer data type to allow dynamic updates to "object" data types. But there were performance issues with the original implementation of Pointer. This release removes many of those performance problems by modifying the evaluator to allow primitive functions to deal with the Pointer type directly, instead of attempting to convert values before passing them to primitive functions.
+
+changes to the haskell API:
+- Introduced a new CustFunc type that should be used to define your own functions when using the Haskell API. This negates having to handle Pointer types in your Haskell code. If you know what you are doing, though, you can handle Pointer types and have a more efficient function
+- Moved runIOThrows to Core, and removed obsolete functions trapError and extractValue
+
 v3.7
 --------